% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_analizar_cuantitativas.R
\name{ds_analizar_cuantitativas}
\alias{ds_analizar_cuantitativas}
\title{Analiza variables cuantitativas.}
\usage{
ds_analizar_cuantitativas(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos}
}
\value{
Retorna la matrix de correlaciones y crea un nuevo
archivo llamado "datos_NoCorrelacionados.csv" con
una de las variables correlaciondas eliminad.
}
\description{
Analiza variables cuantitativas.
}
