% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_balancear_varobjetivo.R
\name{ds_balancear_varobjetivo}
\alias{ds_balancear_varobjetivo}
\title{/usr/bin/Rscript
Balancear la variable objetivo.}
\usage{
ds_balancear_varobjetivo(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
No retorna nada. Modifica los datos con la cantidad
balanceada de Exito y Fracaso de la variable objetivo.
Adicionalmente, crea los histogramas de balanceo.
}
\description{
/usr/bin/Rscript
Balancear la variable objetivo.
}
