% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_binarizar}
\alias{ds_binarizar}
\title{Binariza la variable objetivo a 1: Exito y 0: Fracaso}
\usage{
ds_binarizar(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos sin tranformar.}
}
\value{
datos transformados.
}
\description{
Binariza la variable objetivo a 1: Exito y 0: Fracaso
}
