% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ds_datos_evaluacion}
\alias{ds_datos_evaluacion}
\title{Crea un archivo para evaluacion del modelo}
\usage{
ds_datos_evaluacion()
}
\arguments{
\item{archivoDatos}{Nombre del archivo de donde se obtendrán los datos de prueba.}
}
\description{
Crea un archivo para evaluacion del modelo
}
