% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ds_datos_finales}
\alias{ds_datos_finales}
\title{Retorna y escribe en un archivo los datos finales de desistimiento.}
\usage{
ds_datos_finales()
}
\description{
Retorna y escribe en un archivo los datos finales de desistimiento.
}
