% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ds_datosFormateados}
\alias{ds_datosFormateados}
\title{Retorna y escribe en un archivo los datos formateados de desistimiento.
Estos datos todavía no están preparados para el análisis pero ya se pueden leer en R.}
\usage{
ds_datosFormateados()
}
\description{
Retorna y escribe en un archivo los datos formateados de desistimiento.
Estos datos todavía no están preparados para el análisis pero ya se pueden leer en R.
}
