% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ds_datosPreparados}
\alias{ds_datosPreparados}
\title{Retorna y escribe en un archio los datos preparados de desistimiento.
Estos datos ya se han limpiado y corregido y est치n listos para los an치lisis.}
\usage{
ds_datosPreparados()
}
\description{
Retorna y escribe en un archio los datos preparados de desistimiento.
Estos datos ya se han limpiado y corregido y est치n listos para los an치lisis.
}
