% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_distribuciones.R
\name{ds_distribuciones}
\alias{ds_distribuciones}
\title{Graficar las distribuciones de las variables cualitativas.}
\usage{
ds_distribuciones(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
No retorna nada sino que crea un archivo con los gráficos
de las distribuciones llamado "graficos_distribuciones.png"
}
\description{
Graficar las distribuciones de las variables cualitativas.
}
