% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dividir_datos.R
\name{ds_dividir_datos}
\alias{ds_dividir_datos}
\title{Balancear la variable objetivo.}
\usage{
ds_dividir_datos(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
No retorna nada. Crea dos archivos: uno con los datos
entrenamiento y otro con los datos de pruebas.
}
\description{
Balancear la variable objetivo.
}
