% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_eliminar_variables.R
\name{ds_eliminar_variables}
\alias{ds_eliminar_variables}
\title{Elimina variables de un conjunto de datos.}
\usage{
ds_eliminar_variables(archivoDatos, variables)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}

\item{variables}{Arreglo de nombres de variables para eliminar.}
}
\value{
Retorna una tabla que compara las variables antes y despues.
y guarda los datos a un archivo llamado "datos_Filtrado.csv".
}
\description{
Elimina variables de un conjunto de datos.
}
