% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_entrenar_modelo}
\alias{ds_entrenar_modelo}
\title{Entrena un modelo de regresi贸n l贸gica}
\usage{
ds_entrenar_modelo(archivoDatos, WARNINGS = F)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\description{
Entrena un modelo de regresi贸n l贸gica
}
