% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_evaluar_modelo}
\alias{ds_evaluar_modelo}
\title{Evalúa diferentes métricas a un modelo.}
\usage{
ds_evaluar_modelo(modelo, archivoDatosPruebas)
}
\description{
Evalúa diferentes métricas a un modelo.
}
