% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_formula_modelo}
\alias{ds_formula_modelo}
\title{Retorna la formula del modelo a ajustar deacuerdo a los datos de entrada.}
\usage{
ds_formula_modelo(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\description{
Retorna la formula del modelo a ajustar deacuerdo a los datos de entrada.
}
