% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_guardar_modelo}
\alias{ds_guardar_modelo}
\title{Guarda el modelo de predicci贸n en un archivo "modelo.rds"}
\usage{
ds_guardar_modelo(modelo)
}
\arguments{
\item{modelo:}{Modelo de regresi贸n l贸gica a guardar en el archivo "modelo.rds".}
}
\description{
Guarda el modelo de predicci贸n en un archivo "modelo.rds"
}
