% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_balancear_varobjetivo.R
\name{ds_limpiar_observaciones}
\alias{ds_limpiar_observaciones}
\title{Elimina observaciones con muy pocos datos en categorías (0.5\% del total de los datos).}
\usage{
ds_limpiar_observaciones(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
Nada. Guarda los datos en un nuevo archivo: "datos_Limpios.csv"
}
\description{
Elimina observaciones con muy pocos datos en categorías (0.5\% del total de los datos).
}
