% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_estructura.R
\name{ds_modificar_estructura}
\alias{ds_modificar_estructura}
\title{Modifica la estructura de los datos}
\usage{
ds_modificar_estructura(archivoDatos, archivoEstructura)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos a cambia}

\item{archivoEstructura}{Nombre del archivo con la estructura}
}
\value{
No retorna nada sino que escribe los cambios a un nuevo
archivo de datos llamado "datos_MODIFICADOS.csv"
}
\description{
Modifica la estructura de un conjunto de datos de acuerdo a una
estructura dada como entrada
}
