% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_estructura.R
\name{ds_mostrar_estructura}
\alias{ds_mostrar_estructura}
\title{Imprime y crea un archivo con la estructura actual de los datos del archivo de entrada}
\usage{
ds_mostrar_estructura(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre de archivo con los datos (observaciones X variables).}
}
\description{
Imprime y crea un archivo con la estructura actual de los datos del archivo de entrada
}
