% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ds_obtenerDatos}
\alias{ds_obtenerDatos}
\title{Obtiene los conjuntos de datos}
\usage{
ds_obtenerDatos(tipo = "preparado")
}
\arguments{
\item{tipo}{Cadena de texto que indica el tipo de datos a obtener. Dos posibles valores: "preparado" (por defecto) obtiene los datos ya corregidos y preparados para iniciar los análisis, y "original", obtiene los datos originales del problema.}
}
\value{
Retorna un dataframe con los datos solicitados y además escribe los archivos en el directorio de trabajo en formato CSV (.csv) y Excel (.xlsx)
}
\description{
Obtiene los conjuntos de datos relacionados con desistimiento
}
