% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_predecir}
\alias{ds_predecir}
\title{Predice la variable objetivo "SFinal".}
\usage{
ds_predecir(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo de datos con la
información de las 8 variables para cada individuo.}

\item{modelo}{Modelo cargado o entrenado a utilizar.}
}
\value{
Retorna el mismo conjunto de datos adicionado al
inicio la columna "SFinal" con los valores predichos
de Exito o FRACASO.
}
\description{
Realiza la predicción de la variable objetivo "SFinal"
a partir de los datos de ocho variables: DesembolsoBIE,
PoseeServicioSocial, OcupacionEconomica, EstadoISUN,
NivelEducativo, ExGrupo, ServiciosPúblicos, y TipoVivienda
}
