% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_probar_modelo}
\alias{ds_probar_modelo}
\title{Prueba un modelo de regresi贸n l贸gica.}
\usage{
ds_probar_modelo(modelo, archivoDatos, WARNINGS = F)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\description{
Prueba un modelo de regresi贸n l贸gica.
}
