% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_procesar_nulos.R
\name{ds_procesar_nulos}
\alias{ds_procesar_nulos}
\title{Muestra y procesa los valores nulos}
\usage{
ds_procesar_nulos(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos a procesar.}
}
\value{
Retorna y escribe los resultados a ul archivo "datos_noNulos.csv".
}
\description{
Muesta el numero de nulos por variable de los datos de entrada.
}
