% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_procesar_varobjetivo.R
\name{ds_procesar_varobjetivo}
\alias{ds_procesar_varobjetivo}
\title{Procesa la variable objetivo}
\usage{
ds_procesar_varobjetivo(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
Retorma y escribe los nuevos datos al archivo "datos_Preparados.csv"
}
\description{
Se cambian los nombres y se toman solo los valores de "EXITO" ("Culminado")
y "FRACASO" ("Fuera del Proceso")
}
