% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_resumen_modelo}
\alias{ds_resumen_modelo}
\title{Balancear la variable objetivo.}
\usage{
ds_resumen_modelo(archivoDatos)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\value{
No retorna nada solo retorna una tabla con la variable
objetivo y las variables predictoras.
}
\description{
Balancear la variable objetivo.
}
