% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_modelamiento.R
\name{ds_seleccionar_variables}
\alias{ds_seleccionar_variables}
\title{Selecciona las variables más significativas de un modelo de Regresión Logística.}
\usage{
ds_seleccionar_variables(
  archivoDatosEntrenamiento,
  archivoDatosPruebas,
  WARNINGS = F
)
}
\arguments{
\item{archivoDatos}{Nombre del archivo con los datos.}
}
\description{
Selecciona las variables más significativas de un modelo de Regresión Logística.
}
