% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_datos.R
\name{ver}
\alias{ver}
\title{Visualiza rápidamente la estructura de los datos}
\usage{
ver(datos, n = 5, m = 6, suffix = "")
}
\arguments{
\item{datos}{Conjunto de datos a visualizar}
}
\description{
Visualiza rápidamente la estructura de los datos
}
